requirements.txt
langchain_google_genai
google.generativeai
python-dotenv
langchain_core
streamlit
langchain_community
langserve